requirements.txt
patchify
ipykernel
seaborn
segmentation-models-pytorch
opencv-python
torchmetrics
